EVG-19151: Generate unique webhook secrets
