+++ src/utils/numbers/numbers.test.ts
@@ -1,4 +1,4 @@
-import { roundDecimal } from ".";
+import { roundDecimal, cryptoRandom } from ".";
 
 describe("roundDecimal", () => {
   it("correctly rounds a decimal to the specified number of places", () => {
@@ -9,5 +9,14 @@ describe("roundDecimal", () => {
   it("does not add trailing zeroes", () => {
     expect(roundDecimal(11, 3)).toBe(11);
     expect(roundDecimal(4.2, 3)).toBe(4.2);
+  });
+});
+
+describe("cryptoRandom", () => {
+  it("returns a number between 0 (inclusive) and 1 (exclusive)", () => {
+    const randomNumber = cryptoRandom();
+    expect(typeof randomNumber).toBe("number");
+    expect(randomNumber).toBeGreaterThanOrEqual(0);
+    expect(randomNumber).toBeLessThan(1);
   });
 });
